requirements.txt
# 核心依赖
torch>=1.12.0
numpy>=1.21.0
scipy>=1.7.0
pillow>=8.0.0
pyyaml>=6.0
easydict>=1.9
tqdm>=4.62.0
open3d>=0.15.0
opencv-python>=4.5.0

# 感知损失
piq>=0.7.0

# 分布式训练
mpi4py>=3.1.0

# 开发依赖
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.960
